Use ".so" instead of "lib"
"bin" and "lib" have quite similar shapes, and with a longish list it
turns out to be quite difficult to spot the difference without directly
reading the WHY column.

"bin" and ".so" meanwhile are so distinct it stands out just by scanning
the COMMAND column.  Indeed, I can now spot an out of date library in
use literally from across the room.